+++ sdk/lib/crt/math/libm_sse2/ceil.c
@@ -40,8 +40,8 @@ THE SOFTWARE.
 double FN_PROTOTYPE(ceil)(double x)
 {
   double r;
-  long rexp, xneg;
-  unsigned long ux, ax, ur, mask;
+  long long rexp, xneg;
+  unsigned long long ux, ax, ur, mask;
 
   GET_BITS_DP64(x, ux);
   ax = ux & (~SIGNBIT_DP64);
